Sampling Workflow
Added portal_properties/bika_properties

Added bika_properties/sampling_workflow_enabled (boolean): If
it's False, the "Sampled" transition isn't available.

sample.SamplingDate -> the predicted or scheduled date on which the sample will be taken

sample.DateSampled -> The date the sample was actually "Sampled" in the system.  Defaults to SamplingDate if sampling workflow is disabled.