Made to work on Fedora and behind a proxy.
Added proxy settings for yum/wget/git.  yum and wget uses the http_* environment variables.  git should have done the same, but it misbehaves in a sudo environment on Fedora.  Safest to just set the git proxy if you are behind a firewall/proxy.  Didn't think it is necessary to unset it again if it wasn't set.  The proxy is more than likely permanent in your environment.