Prevent error on AR Reports list when no blob found
If there is a results file without an available blob, the system
throws a POSKeyError: 'No blob file'. This can be caused due to
restore of corrupted blobstorage folders.
This fix prevents the exception being rised and shows the record
under the title 'No available', without the possiblity of being
downloaded. At least, the user can check if sometime ago, a report
was produced for that AR.