Automatic creation of ReferenceAnalyses (Calibration Tests) due to instrument results import
Specs. extract: Some of the results in the instrument file must be parsed and saved automatically
as 'Calibration regular tests' in the system. An instrument picklist selector must be added
to the import form. The calibration results found in the file will be assigned to that instrument.
These objects will be shown on the 'Internal Calibration Test Results View'.