Add Internal Calibration Tests tab inside Instrument View
3004 Bika LIMS upgrade
The table only shows the Reference Analyses (Control and Blanks), the rest
of regular and duplicate analyses linked to this instrument are not displayed.
The Reference Analyses from an Instrument can be from Worksheets (QC analysis
performed regularly for any Analysis Request) or attached directly to the
instrument, without being linked to any Worksheet). In this case, the Reference
Analyses are created automatically by the instrument import tool.