Improvements in results importer tool
Allow to import more than one result for the same sample and SampleID if analysis is
tagged with a Timestamp for ft120 and winescan auto interfaces. Allow to import
different calibration tests for the same QC, Instrument and Analysis, but with
different Capture Dates. Exclude partitions from being saved and use Mean values from
each test instead if exists. Also, add Standard Deviation values if found.
Set the capturedate of the analysis imported to the datetime
defined in the results file.