Allow updating and transitioning multiple items with a single request
Multiple updates can be executed in a single http request, by using the
'update_many' method.  This is a wrapper around the update method.  This
function takes one parameter, called 'input_values', which is a json-encoded
dictionary.  Each key is an obj_path, and each value is a dictionary containing
key/value pairs to be set on the object.  The following input_values parameter
will update two results:

    input_values={"/Plone/clients/client-5/BAR-0014-R01/DM": {"Result": 57},
                  "/Plone/clients/client-5/BAR-0014-R01/CaCO3": {"Result": 76},
                  "/Plone/clients/client-5/BAR-0014-R01/Moist": {"Result": 83}}

The result of the request will be a list of return values from the update method.

It is also possible to transition multiple objects by their object paths, using
a method called doActionFor_many.  This method requires two parameters:

    - obj_paths: A json encoded list of objects to transition.
    - action: The workflow transition to apply to objects.