Cleanup
Demote runtimeerror to warning

WidgetVisibility checks are done at the point where the adapter is
used, not in the adapter itself.  This means that subclassed adapters
don't need to worry about this stuff.

review_states modified centrally in bika_listing.  This means that
subclassed bika_listing/folderitems does not need to worry about this.