Worksheet Management/Edition restrictions in Bika Setup
Requires 3036 Upgrade.

Added two checboxes in BikaSetup > Security:
- Allow access to worksheets only to assigned analysts
- Only lab managers can create new worksheets

When BikaSetup gets modified, changes to ManageWorksheet Permission
mappings are applied thanks to a new bikasetup subscriber class.

IMPORTANT NOTE:
The upgrade set the above options to true by default, so after being
upgraded, only the labmanagers will be able to manage WS and the
analysts will only have access to the worksheets to which they are
assigned.