[LIMS-1737] Fixed error in adding pricelists
If a lab product had no unit and volume, creating a pricelist for lab products gave an error that 'print_detail' was not referenced. This was fixed by initializing print_detail and shifting a case accordingly.

The modified version had been manually tested and creating pricelist of lab products with and without volume and unit was successful.