[LIMS-1724] Fixed dates on reports
In reports, instead of start dates and end dates, '${start_date}' and '${end_date}' are being shown. This is because of a mistake in mapping.

The issue is fixed and is manually tested by generating report of received samples. The report now shows both start dates and end dates or any of them individually.