Performance fix. By default, don't display rejected neither verified WS
This applies for the "All" subview. Since this is the default view, it
takes a lot of time to load when the number of WS is high. If the user
wants to take a look at 'closed' WS (those with verified and/or rejected
statuses), there is already a button for that.