Result below min (or above max) is wrongly replaced in input field
When using specs' '<Min' or '>Max' fields and the result is out of
range, it is replaced by the value entered in the '<Min' or '>Max'
field. This is how it should work, but only in read-only mode.
Otherwise (edite-mode), the original result is lost.

The piece of code removed from inside AjaxCalculateAnalysisEntry fixes
this problem, because this service is only called (used) during edition
of results via javascript. When the results have to be displayed in
read-only mode, the logic to manage these behavior lives inside the .py
codebehind the view, so the replacement in that case works successfully.