Fix TestHiddenAnalyses unit test
This fix is required because the resolution of LIMS-1627 involved the
substution of AR's Profile field to Profiles. This change allowed multiple
Profiles per AR. In fact, the key 'Profile' is no longer valid since
then and when creating an AR manually, 'Profiles' key must be used instead.

The test was not updated when LIMS-1627 was pushed, so it was still using
'Profile'.