Prevent the user to select the method "Not defined" if no manual entry
If manual entry is false for this analysis service, we can assume that only
instrument entry of results is allowed, so we don't want the user to be
able to select a "Non defined" method. His only option is to select an
instrument first. This prevents the user to introduce a result for an
analysis for which only instrument entry is allowed and only has one instrument
assigned which, at the moment, is invalid.