Prevent error when applying a WS Template without a Reference Definitions set
The system allows to create new WS Templates also if there isn't any
Reference Definition registered in the system. When creating a WS by
using the WS Template, the system fails because of this. This quick fix
prevents this to happen, but would be better to do not allow the user
to create Worksheet Templates if there isn't Reference Definitions available