Add UIDReferenceField
This is a very simple uid-based reference field which stores a UID
or a list of UIDS as strings in the zodb, and reconstitutes them
to objects when retrieved.

It does not support backreferences, or link integrity (holdingreference)
but it also does not require the overhead of AT References which is
prohibitive for our use-case.