Assign SampleType's uid instead of its name when sample is set in AR Add
In the funciton sample_set(arnum), the js calls to secondary_ar_sample_info
and it returns a dictionary (data) with the default values that must be
applied to the current Analysis Request. For some fields, there are duplicated
keys for the same object (<fieldname>_uid and <fieldname>), so the values are
not set correctly in the current state_set variable, causing that, for example,
when the system tries to get the information of partitions based on the Sample
Type, the system uses the SampleType name instead of its uid, making the system
fail to return that partition information.