Fix error when submitting duplicate analyses
The error was caused by getSiblings function from DuplicateAnalysis,
cause it was retrieving all analyses associated to the worksheet w/o
discarding those that are reference analyses (that do not have the
function getRequestUID). getSiblings method is used to obtain other
analyses that are closely related with the current one (e.g. to know
which are the dependencies/dependants of the current analysis). For
DuplicateAnalysis, getSiblings method returns all the analyses that
are associated to the same analysis request, duplicates included.

Whit this commit, the new IRequestAnalysis has been created. This
interface is used to distinguish between analyses that have a request
assigned from those that do not have one. Thus, this interface
provides the signatures for methods related with the retrival of the
request associated to the analysis as well as main related info (UID,
ID, URL, Client, etc.)