Passing doctest Permissions.rst
Some changes were made from the original in bikalims/bika.lims:master
but they are all reasonable changes.  There were a few changes that were
made to the senaite-integration from master that I do not agree with,
but we have roadmapped a plan to fix these soon in wip branch.
certainly senaite-integration is not the correct branch for fixing the
default workflow permissions.