Show outofrange icons in analyses views
Out of range icons were not displayed in analyses views and were only
working on results input cause in that scenario, the rendering of
out-of-range icons was managed through js.

The rendering of out-of-range icons was lost when AnalysesView was
refactored to improve its performance. Before then, bikalisting base
view was responsible of rendering those icons through the "classic"
folderitems function.