Clarification about IFieldIcons in bikalisting
In classic/old folderitems function from bikalisting base class,
the system was always looking for IFieldIcons adapters for the
object to be rendered in order to fill self.field_icons var,
which is used in the template to display additional icons for
the object's counterpart row representation. These adapters are
only used in Analyses, so there was no reason to keep them in the
newer implementation of folder_items (which uses brains by default
instead of full objects).