Out of range Icons are not displayed through all Analysis states (#603)
* Show outofrange icons in analyses views

Out of range icons were not displayed in analyses views and were only
working on results input cause in that scenario, the rendering of
out-of-range icons was managed through js.

The rendering of out-of-range icons was lost when AnalysesView was
refactored to improve its performance. Before then, bikalisting base
view was responsible of rendering those icons through the "classic"
folderitems function.

* Clarification about IFieldIcons in bikalisting

In classic/old folderitems function from bikalisting base class,
the system was always looking for IFieldIcons adapters for the
object to be rendered in order to fill self.field_icons var,
which is used in the template to display additional icons for
the object's counterpart row representation. These adapters are
only used in Analyses, so there was no reason to keep them in the
newer implementation of folder_items (which uses brains by default
instead of full objects).

* Changelog #603