Analysis editable if not submitted, regardless of point of capture
One of the conditions for analysis to be edited is that a result mustn't
be submitted already. This check was done when the point of capture is
"field" (the result can be set when the sample is collected), but this
condition was not checked if the point of capture is "lab" (the result
must be set in the lab, after the sample is received).