Change index 'WorksheetTemplate' to 'WorksheetTemplateTitle'
`bika_catalog_worksheet_listing` is used to store the worksheets. This
catalog had a FieldIndex `WorksheetTemplate`, so when reindexing
worksheets with a `WorksheetTemplate` associated, the system was throwing
the following Traceback:

TypeError: Can't pickle objects in acquisition wrappers.

The index name has been renamed to `WorksheetTemplateTitle`, cause
`Worksheet` objects have a function that returns directly the title
of the `WorksheetTemplate` associated.