Assign/unassign AR automatically depending on its analyses status
When an analysis is transitioned to `assigned` state, the system must
transition the parent Analysis Request to `assigned` state too, but only
if all the analyses it contains are in `assigned` already.

When an analysis is transitioned to `unassigned` state, the system must
transition the parent Analysis Request automatically to `unassigned` state
too.