Better management of range shoulders with warn_min and warn_max
The error percentage for ranges was meaningless. The calculation of
shoulders was done by applying the %error to the result, regardless
of the scale or coverage of all the valid range. Applying a %error
based on the whole length of the valid range was not acceptable
neither, cause the range could be logarithmic (e.g. pH) or the
difference between min and max could be too wide, result on
meaningless shoulders, mostly on the left side.