Make DuplicateAnalysis' getResultsRange consistent
Since DuplicateAnalysis inherits from AbstractAnalysis, getResultsRange
was returning the range assigned to the original analysis (acquired from
the corresponding Analysis Service). This is wrong, a Dupliucate must be
considered out of range if its result does not match with the result set
to the original analysis, with a duplicate variation in % as the margin
error. Thus, with this commit, getResultsRange from DuplicateAnalysis
overrides its parent function and returns a dictionary where min and max
are the original analysis result +/- the duplicate variation.