Make getResultsRange function from Reference Analysis consistent
getResultRange was returning the results range dict from the Reference
Sample the Reference Analysis refers to, so it was a dictionary with
service uids as keys and the resultrange for each service as the value.
Routine Analyses and Duplicate analyses are returning the dictionary
with the keys min, max, etc. that apply especifically to them.
Thus, the function has been revisited so with this commit, it returns
only the result range that applies to itself, that is the discrete
value set in parent's Sample Reference +/- %error