Remove specifications paramter from getResultsRange functions
This paramter is confusing and is not used anywhere. Analysis Request's
getResultsRange already deals with the specification that was assigned
to the Analysis Request (and thus, to the Analyses it contains) either
during AR Add or in Manage Analyses. The user was responsible to choose
an AnalysisSpecification from lab or from client at that level, so there
is no need to sweep this paramter along everywhere.