Remove remaining 'error' subfield from specifications
As per the previous commits, error subfield is only used when specs for Reference
Definitions/Samples are submitted (and only during the submission process) for the
calculation of `min` and `max` subfields. Error field is no longer used anywhere in
favour of `warn_min` and `warn_max` subfields, that allow the user to explicitily
set the boundaries of the range without the caveats of using a % for a range.