Fix ARAnalysesField.rst
There is no precedence of Specifications by default in Analysis Requests.
Specifications must be set especifically to the Analysis Request in AR Add
form. The system used to resolve the specification automatically, which is
totally wrong, cause this decision must be done by lab manager. With the
latest code, if the user does not choose any specification, the system will
not use (or try to resolve) any. Thus, this use case from the doctest is
not valid anymore