Prevent uidreferencefield from extra queries for brains
is_uid function calls to uid_catalog directly to see if the uid is valid.
get_object calls this function first to check if the uid is valid, but
do a search against the catalog to obtain the object. With this commit,
we use api's get_uid (that does not validate against uid_catalog by default)
to check the uid, so only one query is done at the end.