Analyses submission in Worksheet is slow (#776)
* Prevent uidreferencefield from extra queries for brains

is_uid function calls to uid_catalog directly to see if the uid is valid.
get_object calls this function first to check if the uid is valid, but
do a search against the catalog to obtain the object. With this commit,
we use api's get_uid (that does not validate against uid_catalog by default)
to check the uid, so only one query is done at the end.

* Overhead redux in workflow's doAction

When doAction is not possible, the system was obtaining the available
transitions for the given object to print them in the log only. We can
remove this overhead.

* Move the AR auto-submit to the analysis' after_submit event

* Better performance of ws' submit guard

* Slightly better performance in AR's getAnalysesNum

api's is_active works with brains, while isActive from workflow always
look through the workflow chain

* Small bits of performance for ws' analyses submission

* Added #776 in CHANGES.rst

* Remove is_uid function from uidreferencefield

This function is no longer used, api.is_uid is used everywhere instead

* Make _get_object from uidreferencefield to use api's functions

* Revisit worksheet's analyses submission and instrument refactoring

Most of the overhead during analyses submission was because of their
relationship with instruments and reference analysis validity. This
commit lightens content.instrument and makes the submission of
analyses more optimal.

* Sanitize analysis submission in ARs

* Added upgrade step to remove InstrumentLatestReferenceAnalyses reference

* New package browser.analyses

Moved browser.analyses.py and browser.late_analyses.py inside

* Remove duplicated code and unify analyses submission in lists

* Fix CannotGetPortalError: Unable to get the portal object

Revert _get_object changes in uidreferencefield. In that situation,
the context is not available