Check if the analysis has a partition assigned before removing refs
At this point, might happen the analysis to be removed does not have
does not have a partition assigned already, precisely because of the
inherent recursive behavior caused by lines L206 + L208.
Even though, this is temporary and will only happen during the
process of assigning/unassigning analyses here inside this set func.