Display Analysis Services sorted ascending
When expanding Analysis Categories, an ajax_call is submitted to get
only the tr. In POST, the `sort_order` is passed in, but empty, so
the system assumes its value must be "descending" by default. With
this commit, if an empty value is passed in, the sytem uses the default
sort_order defined in the view.