Rebind retracted ARs to ParentRetracted field
The field "ParentAnalysisRequest" from AnalysisRequest was used to
keep the relationship between retracted ARs and retests (ARs). The
name of this field seems more appropiate for storing the relationship
amongst partitions and the analysis requests they are originated from.

Thus, for storing the retracted-retest relationship, the new field
"ParentRetracted" will be used.