Display partitions that get orphan because of search/pagination
Analysis request listing may return partitions, but without their
parent. In this case, the listing was simply omitting them. Hence,
if the user searches for e.g. a partition id, the list does not
display any result.

This commit displays the partition instead of omitting it, but
applies a particular css class "orphan-partition" so at least, is
displayed differently from those that are not orphan.