Enhanced partitioning system (Template + partition magic) (#1068)
* Added guard_handler to make guards easier to mantain

With this approach, there is only the need to set
`here.guard_handler(<transition_id>)` in DC Workflow definition and
senaite will look for a `guard_<transition_id>` function in
workflow.<portal_type>.guards

* Added create_partitions transition and guard

create_partition transition can only be triggered by lab personnel (labclerk,
labman, analyst) when the analysis request state is sample_received. Also,
we assume here that the transition can only be performed if the partitioning
is enabled in Setup (for now, field ShowPartitions) and the analysis request
is active.

* Upgrade to add create_partition transition and guard

* Rebind retracted ARs to ParentRetracted field

The field "ParentAnalysisRequest" from AnalysisRequest was used to
keep the relationship between retracted ARs and retests (ARs). The
name of this field seems more appropiate for storing the relationship
amongst partitions and the analysis requests they are originated from.

Thus, for storing the retracted-retest relationship, the new field
"ParentRetracted" will be used.

* Refactor relationships between retracted and retested ARs

Field 'Retest' instead of 'ChildAnalysisRequest'
Field 'Retracted' instead of 'ParentAnalysisRequest'

* Remove the orphan relationship "ChildAnalysisRequest"

* Add getAncestorsUIDs index for partitioning

* Masking of Analyses in searches by Analysis Request

* Add supermodel decorator in senaite.core

* Add SampleType field in Template's partitions view

* Add partition magic view

* Added create_partition transition in lists

* Remove PrimarySample assignment on partition creation

* Function comment

* Display partitions in AR list indented and below their parent

* Fix KeyError: table_row_class

* Update CHANGES.rst

* Display partitions that get orphan because of search/pagination

Analysis request listing may return partitions, but without their
parent. In this case, the listing was simply omitting them. Hence,
if the user searches for e.g. a partition id, the list does not
display any result.

This commit displays the partition instead of omitting it, but
applies a particular css class "orphan-partition" so at least, is
displayed differently from those that are not orphan.

* Set default page size in listings to 50 items

* Do not allow the creation of partitions from other partitions