Guards submit and attach for AR must consider masked analyses
The criteria must be based on the analyses that belong to the AR, but
need to consider the masking because of partitioning. The AR contains
its own analyses, but as soon as an AR-alike partition is created,
the analyses from the partition mask the parents, which will no longer
be used.