Make Parent ARs to follow their partitions in workflow
When all the partitions from an AR has been submitted (their status
is `to_be_verified`), the parent AR must transition automatically
too. The same principle applies for the rest of transitions ("verify",
"publish", "cancel" and "reinstate").

With the same reasoning, when the Parent AR is transitioned to
"published", "cancelled", etc., the partitions must follow.