Make Parent ARs to follow their partitions in workflow (#1074)
* ARAnalysesField refactoring

To make partitioning work, the query to fetch the analyses from the
current object (AR) must use `getRequesUID` parameter in order to
get all masked analyses from partitions

* ARAnalysesField.get redux (removal of "retracted" param)

"retracted" param in kwargs is not used anywhere except in `getLate`
function. Hence, better to keep the function less complex and just
add the "retracted" check in `getLate` function.

* Guards submit and attach for AR must consider masked analyses

The criteria must be based on the analyses that belong to the AR, but
need to consider the masking because of partitioning. The AR contains
its own analyses, but as soon as an AR-alike partition is created,
the analyses from the partition mask the parents, which will no longer
be used.

* Make Parent ARs to follow their partitions in workflow

When all the partitions from an AR has been submitted (their status
is `to_be_verified`), the parent AR must transition automatically
too. The same principle applies for the rest of transitions ("verify",
"publish", "cancel" and "reinstate").

With the same reasoning, when the Parent AR is transitioned to
"published", "cancelled", etc., the partitions must follow.