Remove superfluous "assign" transition from ref/dup analyses
Since reference analyses and duplicates can only exist if they are
bound to a Worksheet (and/or to an Instrument for reference analyses),
make the initial state to be assigned. For the same reason, an
"unassigned" reference analysis or duplicate will not be accessible
anywhere anymore, so better to make the "unassign" transition to remove
these analyses from the system. Note that "unassign" transition is
only permitted when the state is "assigned", so there are no results
yet and the removal is safe.

This commit includes some refactoring, improvements and redux of
content.worksheet to make the addition/removal of analyses (and their
counterpart transitions "assign" and "unassign") faster.