Remove Retested field from Analysis - RetestedOf instead
Retested field was confusing, if the value for this field was True,
it meant the current analysis was a retest, not the other way round.
This field has been removed and an additional field "RetestedOf" has
been added, as well as a function "isRetest".

Index "getRetested" (FieldIndex) from analysis_catalog has been
replaced by "isRetested" (BooleanIndex). Metadata column "getRetested"
has been replaced by "getRetestedOfUID", that returns the UID of the
retracted analysis the current analysis is a retest of.