Selective reindexing of indexes in transitions (#1100)
There is no need to reindex all indexes after a transition is performed
for a given object. In fact, in most cases there's only the need to
reindex two or three indexes. Reindexing only those indexes that might
be affected by the transition boosts the overall performance.

With this commit, the system uses the ACTION_INDEXES mapping (located at
`workflow.indexes`) to infere the indexes that need to be reindexed for
a given portal_type after a transition is performed. If no indexes are
found in the mapping, the system behaves as usual, reindexing all the
indexes of the object.