Add AR's rollback to receive for retract transition
By default, senaite does not allow the retraction of verified
analyses, but is quite common this restriction to be removed
in custom add-ons. This commit adds the rollback by default
when the transition takes place, so only adding the exit
transition `rollback_to_receive` for state `verified` in custom
add-ons should be enough for this to work.