Bound Analysis Request fields to field-specific permission in wf
There is no longer need to hardcode the visibility for every single
state and field in the AnalysisRequest content type. The system uses
now the `write_permission` and `read_permission` attributes to know
if a given field must be rendered or not depending on the current
status of the Analysis Request.

If not explicitily set, the system assumes the visibility for "view"
and "edit" modes is "visible", so there is no need to define these
modes in the field itself unless we want to not render that field in
that mode regardless of permissions.

This commit also improves (and fix some issues) the adapters for
widgetvisibility. Although the primary criteria is to use field-
specific permissions in workflow, sometimes we need more fine-grained
behavior (we want a field to be displayed only if a given setting
from setup is enabled, or make editable a field when other field has
value, etc.) that can be achieved by making use of these adapters.

There are still some fields from AR that need to be ported and maybe
some additional bits of improvement, but this commit comes with the
basics in place.