Fix AnalysisRequest.rst doc test
Is not responsibility of the adapters to check permissions. Permissions
are checked directly by Zope and workflow machinery by making use of
`write_permission` and `read_permission` fields. If an adapter
wrongly changes a field widget from invisible to visible in edit mode
(and the permission is not granted), the system will not allow the user
to store a value for that field anyways. It makes the system more robust