Display the preservation, even if preservation wf is disabled
Preservation workflow is meant to be used when the sampling wf is
enabled and is necessary to a) preserve the sample, b) know who
preserved the sample and c) when the sample was preserved.
Nevertheless, capture of field preservation is still important in
basic workflow (with preservation disabled), cause is used by labclerk
to check, together with the container and sample type if the sample
arrived in good conditions and can be accepted.