Add "Auto-partition" bool field in Template
By default, when a Template is used for Sample creation, the system
redirects to the partitioning view when the Sample is received.
Nevertheless, some laboratories do not perform the partitioning at
the same time the sample is received, rather they do it in a separate
process, either because the user in charge of reception does not have
the skills for the partitioning or because they sent the sample to a
department that is specifically in charge of partitioning and storage.

With the changes provided in this commit, the user (labman) can
enable/disable the auto-partitioning on reception in the Template
itself. Hence, when this option is disabled, the user will not be
redirected to the partitions creation view, while the primary sample
will be successfully transitioned to `received` status.