Remove implicit definition of transitions in AR listing
Manual definition of transitions in listings is no longer need, but is also
disencouraged. The rendering of the available transitions is automatically
done, by grabbing the allowed transitions for a given state that are defined
in DC Workflow.

Removing implicit definition of transitions in listings also makes the
development of add-ons easier, cause otherwise, the browser views need to
be overriden to display new transitions not defined in senaite.core.