Do not modify the analysis ID on retraction
The analysis ID was changed on retraction with the aim to keep the
same ID for the retest. So, if the analysis ID was "Mg", after
retraction it was changed to "Mg-1", while "Mg" ID was set to the
retest.

This commit changes this behavior so the ID of the original analysis
is kept, while the ID of the retest is set to "<keyword>-<num>".