Display "unassign" transition in Worksheet context only
The transition "unassign" is currently displayed for analyses listings
in contexts other than Worksheet (e.g. in Sample -AR- context). Although
this is not a problem because for the unassignment of an Analysis from
a worksheet there is no need to know the worksheet the analysis is
assigned to, it might be confusing and could accidentally cause undesired
unassignments.

This commit, makes the unassign transition button to only be displayed
inside worksheet contexts. It also injects the current worksheet uid in
the request when calling "removeAnalysis" from Worksheet context.
Otherwise, wouldn't be possible to unassign the analysis programatically
regardless of the current context.