Better handling of IDs without separator
Is quite common to use IDs without separators, cause it allows to
add more information in less space. This is specially important when
small labels are used. An example of ID formatting without separators
can be: "{year}{sampleType}{alpha:3a2}' for AnalysisRequest type.

Although the system already allows this configuration and next id is
resolved correctly each time a new object is created, number generator
does not behave as expected: instead of splitting and storing the
prefix to make the creation of further IDs faster, number generator
stores the whole string for every single object and system is forced
to always look through the objects to find out the next suitable ID.

For example, for an ID generated using the formatting above:

2019WBAAA01

the value stored in annotation storage is a single entry for the
wole string (prefixed with the portal_type):

`analysisrequest-2019WBAAA01`

If we create a new AR (2019WBAAA02), annotation storage will add a new
entry:

`analysisrequest-2019WBAAA02`

Number generator relies both on the "split_length" and "separator"
params to find out what's the "sequential" part of the ID. Note that
since there is no separator, the result will always be the same,
regardless of "split_length" value.

As a result, this behavior ends with low performance and also a huge
annotations storage for IDs.

This commit fixes this problem by, besides the "separator" split, to
always split the ID by wildcards/keywords. This is, the format above
is splitted like follows, even if no separator is set:

["{year}", "{sampleType}", "{alpha:3a2d}"]

Thus, setting a "split_length" of 2 will generate only one entry in
NumberGeneration storage when the two ARs above are created:

`analysisrequest-2019WB`

And will keep the alphanumeric part as the sequential portion
accordingly.