No need of submit adapter for AR
The submit transition of an AR is done automatically when all the
analyses it contains are submitted. There is a guard that prevents the
transition "submit" to be displayed in ARs unless all analyses are
submitted. Hence, there is no need to have an adapter for "submit"
request action for ARs