Cleanup "Transition: Verify" from ar_workflow and add "Invalidate"
"Transition: Verify" only applies to analysis workflow, not to AR
workflow. The reason is that verify transition takes place automatically
as soon as the transition "verify" is done against all individual analyses
the AR contains. Thus, the transition "verify" from AR ultimately relies
on the guard and premissions set for "verify" in analysis_workflow.xml

On the other hand, permission guard for invalidate transition was mixed
with the permission for retraction. Retraction and invalidation are two
different things.