Harmonize type views
Since there are users that can have View permission for a given type but
not Modify portal content permission, the default view for types (at least
inside setup) should be base_view. Otherwise, if the user has View
permission, but not Modify portal content permission, an Insufficient
Privileges error is displayed, without other information. So, the user
cannot see the contents of the type unless he/she types /view or /base_view
at the end of the url